# frontend/src/api/openapi.json
{
  "components": {
    "schemas": {
...
      "DocumentDetail": {
        "properties": {
          "archive_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Archive Url"
          },
          "arxiv_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Arxiv Id"
          },
          "authors": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Authors",
            "type": "array"
          },
          "chunks_count": {
            "title": "Chunks Count",
            "type": "integer"
          },
          "cluster_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cluster Id"
          },
          "cluster_label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cluster Label"
          },
          "collections": {
            "items": {
              "type": "string"
            },
            "title": "Collections",
            "type": "array"
          },
          "date_added": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Date Added"
          },
          "description": {
            "default": "",
            "title": "Description",
            "type": "string"
          },
          "doi": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Doi"
          },
          "doi_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Doi Url"
          },
          "enrichment": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/EnrichmentOut"
            },
            "title": "Enrichment",
            "type": "array"
          },
          "fetch_status": {
            "title": "Fetch Status",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "image": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ImageDetail"
              },
              {
                "type": "null"
              }
            ]
          },
          "isbn": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Isbn"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "overlay_tags": {
            "items": {
              "$ref": "#/components/schemas/TagOut"
            },
            "title": "Overlay Tags",
            "type": "array"
          },
          "reddit": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RedditDetail"
              },
              {
                "type": "null"
              }
            ]
          },
          "similarity": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Similarity"
          },
          "source": {
            "title": "Source",
            "type": "string"
          },
          "source_id": {
            "title": "Source Id",
            "type": "string"
          },
          "source_tags": {
            "items": {
              "type": "string"
            },
            "title": "Source Tags",
            "type": "array"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "url_or_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Url Or Path"
          },
          "year": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Year"
          },
          "zotero_attachment_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Zotero Attachment Key"
          }
        },
        "required": [
          "id",
          "source",
          "source_id",
          "title",
          "url_or_path",
          "date_added",
          "fetch_status",
          "source_tags",
          "overlay_tags",
          "cluster_id",
          "cluster_label",
          "chunks_count",
          "collections"
        ],
        "title": "DocumentDetail",
        "type": "object"
      },
      "DocumentListItem": {
        "properties": {
          "archive_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Archive Url"
          },
          "cluster_l1_tags": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Cluster L1 Tags",
            "type": "array"
          },
          "cluster_l2_tags": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Cluster L2 Tags",
            "type": "array"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "source": {
            "title": "Source",
            "type": "string"
          },
          "source_id": {
            "title": "Source Id",
            "type": "string"
          },
          "source_tags": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Source Tags",
            "type": "array"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "url_or_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Url Or Path"
          },
          "zotero_attachment_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Zotero Attachment Key"
          }
        },
        "required": [
          "id",
          "source",
          "source_id",
          "title",
          "description"
        ],
        "title": "DocumentListItem",
        "type": "object"
      },
      "DocumentListResponse": {
        "properties": {
          "documents": {
            "items": {
              "$ref": "#/components/schemas/DocumentListItem"
            },
            "title": "Documents",
            "type": "array"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "total",
          "documents"
        ],
        "title": "DocumentListResponse",
        "type": "object"
      },
...
      "EnrichmentOut": {
        "description": "One retrieval-enrichment chunk's provenance (DESIGN.md \u00a73.2).\n\n``label`` is rendered here rather than in the frontend so the enrichment\nladder stays the single source of truth for how each rung is named.",
        "properties": {
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "ref_title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ref Title"
          },
          "resolved_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolved By"
          },
          "source_ref": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Ref"
          },
          "text": {
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "kind",
          "resolved_by",
          "label",
          "source_ref",
          "ref_title",
          "text"
        ],
        "title": "EnrichmentOut",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "ImageDetail": {
        "description": "Image-specific fields, present only when ``source == image``.",
        "properties": {
          "image_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Type"
          },
          "ocr_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ocr Text"
          }
        },
        "title": "ImageDetail",
        "type": "object"
      },
...
      "RedditDetail": {
        "description": "Reddit-specific fields, present only when ``source == reddit``.\n\n``body`` is the post or comment verbatim, not the 280-char ``description``\nthe cards use. ``permalink`` is the reddit thread; for a link post that is a\ndifferent URL from ``url_or_path``, which points at the external target.",
        "properties": {
          "body": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Body"
          },
          "external_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "External Url"
          },
          "kind": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Kind"
          },
          "permalink": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Permalink"
          },
          "subreddit": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subreddit"
          }
        },
        "title": "RedditDetail",
        "type": "object"
      },
...
      "Source": {
        "enum": [
          "firefox",
          "zotero",
          "calibre",
          "image",
          "youtube",
          "reddit"
        ],
        "title": "Source",
        "type": "string"
      },
...
      "TagOut": {
        "properties": {
          "confidence": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confidence"
          },
          "origin": {
            "title": "Origin",
            "type": "string"
          },
          "tag": {
            "title": "Tag",
            "type": "string"
          }
        },
        "required": [
          "tag",
          "origin"
        ],
        "title": "TagOut",
        "type": "object"
      },
      "TagPatchRequest": {
        "properties": {
          "add": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Add",
            "type": "array"
          },
          "remove": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Remove",
            "type": "array"
          }
        },
        "title": "TagPatchRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Local-first research library API",
    "title": "Alexandria",
    "version": "0.0.9"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/clusters/{cluster_id}/documents": {
      "get": {
        "operationId": "cluster_documents_clusters__cluster_id__documents_get",
        "parameters": [
          {
            "in": "path",
            "name": "cluster_id",
            "required": true,
            "schema": {
              "title": "Cluster Id",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cluster Documents",
        "tags": [
          "clusters"
        ]
      }
    },
...
    "/documents": {
      "get": {
        "operationId": "list_documents_documents_get",
        "parameters": [
          {
            "in": "query",
            "name": "sources",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "$ref": "#/components/schemas/Source"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Sources"
            }
          },
          {
            "in": "query",
            "name": "source_tags",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Source Tags"
            }
          },
          {
            "in": "query",
            "name": "general_tags",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "General Tags"
            }
          },
          {
            "in": "query",
            "name": "overlay_tags",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Overlay Tags"
            }
          },
          {
            "in": "query",
            "name": "cluster_l1_tags",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cluster L1 Tags"
            }
          },
          {
            "in": "query",
            "name": "cluster_l2_tags",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cluster L2 Tags"
            }
          },
          {
            "in": "query",
            "name": "learned_tags",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Learned Tags"
            }
          },
          {
            "in": "query",
            "name": "wayback_only",
            "required": false,
            "schema": {
              "default": false,
              "title": "Wayback Only",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 48,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Documents",
        "tags": [
          "documents"
        ]
      }
    },
    "/documents/{doc_id}": {
      "get": {
        "operationId": "get_document_documents__doc_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "doc_id",
            "required": true,
            "schema": {
              "title": "Doc Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Document",
        "tags": [
          "documents"
        ]
      }
    },
    "/documents/{doc_id}/cover": {
      "get": {
        "description": "Serve a document's cover image.\n\nCalibre books have a ``cover.jpg`` next to their format files; image\ndocuments *are* the image, so their ``url_or_path`` is streamed directly.",
        "operationId": "get_document_cover_documents__doc_id__cover_get",
        "parameters": [
          {
            "in": "path",
            "name": "doc_id",
            "required": true,
            "schema": {
              "title": "Doc Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Document Cover",
        "tags": [
          "documents"
        ]
      }
    },
    "/documents/{doc_id}/tags": {
      "patch": {
        "operationId": "patch_tags_documents__doc_id__tags_patch",
        "parameters": [
          {
            "in": "path",
            "name": "doc_id",
            "required": true,
            "schema": {
              "title": "Doc Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TagPatchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Patch Tags Documents  Doc Id  Tags Patch",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Patch Tags",
        "tags": [
          "documents"
        ]
      }
    },
    "/documents/{doc_id}/text": {
      "get": {
        "description": "The body text retained at ingestion, verbatim.\n\nThe audit surface for \"what did the fetcher actually get\" \u2014 the question\nthat could previously only be guessed at from the chunks a document\nproduced. 404 when the document has\nno retained text, which is the ordinary case for anything ingested before\nretention shipped: there is no backfill.",
        "operationId": "get_document_text_documents__doc_id__text_get",
        "parameters": [
          {
            "in": "path",
            "name": "doc_id",
            "required": true,
            "schema": {
              "title": "Doc Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Document Text",
        "tags": [
          "documents"
        ]
      }
    },
...
  }
}